+++ src/extensions/image/index.ts
@@ -100,6 +100,8 @@ const imageResizable = (editor: Editor) => {
           event.interaction.stop()
           return
         }
+        //取消dom监听
+        editor.removeDomObserve()
         //禁用dragstart
         DapEvent.on(event.target, 'dragstart', e => e.preventDefault())
         //获取图片节点