+++ src/extensions/video/index.ts
@@ -102,6 +102,8 @@ const videoResizable = (editor: Editor) => {
           event.interaction.stop()
           return
         }
+        //取消dom监听
+        editor.removeDomObserve()
         //禁用dragstart
         DapEvent.on(event.target, 'dragstart', e => e.preventDefault())
         //获取视频节点